Add XACML DAO Registry Implementation
