+++ components/entitlement/org.wso2.carbon.identity.entitlement/src/main/java/org/wso2/carbon/identity/entitlement/PDPConstants.java
@@ -291,4 +291,18 @@ public class PDPConstants {
 
     public static final String STORE_POLICY_META_DATA = "PAP.Policy.Store.MetaData";
 
+    public static final String GLOBAL_POLICY_COMBINING_ALGORITHM = "globalPolicyCombiningAlgorithm";
+
+    public static final String MAX_NO_OF_STATUS_RECORDS = "maxRecodesToPersist";
+
+    public static final int DEFAULT_MAX_NO_OF_STATUS_RECORDS = 50;
+
+    public static final String MAX_NO_OF_POLICY_VERSIONS = "maxVersions";
+
+    public static final int DEFAULT_MAX_NO_OF_POLICY_VERSIONS = 5;
+
+    public static final String POLICY_STORE_PATH = "policyStorePath";
+
+    public static final String DEFAULT_POLICY_STORE_PATH = "/repository/identity/entitlement/policy/pdp/";
+
 }
