+++ components/entitlement/org.wso2.carbon.identity.entitlement/src/main/java/org/wso2/carbon/identity/entitlement/pap/store/PAPPolicyStoreReader.java
@@ -23,6 +23,8 @@ import org.wso2.balana.AbstractPolicy;
 import org.wso2.balana.finder.PolicyFinder;
 import org.wso2.carbon.identity.entitlement.EntitlementException;
 import org.wso2.carbon.identity.entitlement.PDPConstants;
+import org.wso2.carbon.identity.entitlement.dao.PolicyDAO;
+import org.wso2.carbon.identity.entitlement.dto.AttributeDTO;
 import org.wso2.carbon.identity.entitlement.dto.PolicyDTO;
 import org.wso2.carbon.identity.entitlement.pap.PAPPolicyReader;
 import org.wso2.carbon.identity.entitlement.policy.PolicyAttributeBuilder;
@@ -38,33 +40,29 @@ public class PAPPolicyStoreReader {
     // the optional logger used for error reporting
     private static Log log = LogFactory.getLog(PAPPolicyStoreReader.class);
 
-    private PAPPolicyStore store;
+    private final PolicyDAO store;
 
     /**
      * @param store
      */
-    public PAPPolicyStoreReader(PAPPolicyStore store) {
+    public PAPPolicyStoreReader(PolicyDAO store) {
         this.store = store;
     }
 
 
     /**
-     * @param policyId
-     * @param finder
-     * @return
-     * @throws EntitlementException
+     * @param policyId policyId
+     * @param finder   policy finder
+     * @return abstract policy
+     * @throws EntitlementException, throws, if fails
      */
     public synchronized AbstractPolicy readPolicy(String policyId, PolicyFinder finder)
             throws EntitlementException {
-        Resource resource = store.getPolicy(policyId, PDPConstants.ENTITLEMENT_POLICY_PAP);
-        if (resource != null) {
-            try {
-                String policy = new String((byte[]) resource.getContent(), Charset.forName("UTF-8"));
-                return PAPPolicyReader.getInstance(null).getPolicy(policy);
-            } catch (RegistryException e) {
-                log.error("Error while parsing entitlement policy", e);
-                throw new EntitlementException("Error while loading entitlement policy");
-            }
+
+        PolicyDTO dto = store.getPAPPolicy(policyId);
+        if (dto != null) {
+            String policy = dto.getPolicy();
+            return PAPPolicyReader.getInstance(null).getPolicy(policy);
         }
         return null;
     }
@@ -72,17 +70,13 @@ public class PAPPolicyStoreReader {
     /**
      * Reads All policies as Light Weight PolicyDTO
      *
-     * @return Array of PolicyDTO but don not contains XACML policy and attribute meta data
+     * @return Array of PolicyDTO but does not contain XACML policy and attribute metadata
      * @throws EntitlementException throws, if fails
      */
     public PolicyDTO[] readAllLightPolicyDTOs() throws EntitlementException {
 
-        String[] resources = null;
-        resources = store.getAllPolicyIds();
-
-        if (resources == null) {
-            return new PolicyDTO[0];
-        }
+        String[] resources;
+        resources = store.listPolicyIds().toArray(new String[0]);
 
         List<PolicyDTO> policyDTOList = new ArrayList<PolicyDTO>();
 
@@ -90,8 +84,7 @@ public class PAPPolicyStoreReader {
             PolicyDTO policyDTO = readLightPolicyDTO(resource);
             policyDTOList.add(policyDTO);
         }
-
-        return policyDTOList.toArray(new PolicyDTO[policyDTOList.size()]);
+        return policyDTOList.toArray(new PolicyDTO[0]);
     }
 
     /**
@@ -102,82 +95,27 @@ public class PAPPolicyStoreReader {
      * @throws EntitlementException throws, if fails
      */
     public PolicyDTO readPolicyDTO(String policyId) throws EntitlementException {
-        Resource resource = null;
-        PolicyDTO dto = null;
-        try {
-            resource = store.getPolicy(policyId, PDPConstants.ENTITLEMENT_POLICY_PAP);
-            if (resource == null) {
-                log.error("Policy does not exist in the system with id " + policyId);
-                throw new EntitlementException("Policy does not exist in the system with id " + policyId);
-            }
 
-            dto = new PolicyDTO();
-            dto.setPolicyId(policyId);
-            dto.setPolicy(new String((byte[]) resource.getContent(), Charset.forName("UTF-8")));
-            dto.setActive(Boolean.parseBoolean(resource.getProperty(PDPConstants.ACTIVE_POLICY)));
-            String policyOrder = resource.getProperty(PDPConstants.POLICY_ORDER);
-            if (policyOrder != null) {
-                dto.setPolicyOrder(Integer.parseInt(policyOrder));
-            } else {
-                dto.setPolicyOrder(0);
-            }
-            dto.setPolicyType(resource.getProperty(PDPConstants.POLICY_TYPE));
-            String version = resource.getProperty(PDPConstants.POLICY_VERSION);
-            if (version != null) {
-                dto.setVersion(version);
-            }
-            String lastModifiedTime = resource.getProperty(PDPConstants.LAST_MODIFIED_TIME);
-            if (lastModifiedTime != null) {
-                dto.setLastModifiedTime(lastModifiedTime);
-            }
-            String lastModifiedUser = resource.getProperty(PDPConstants.LAST_MODIFIED_USER);
-            if (lastModifiedUser != null) {
-                dto.setLastModifiedUser(lastModifiedUser);
-            }
-            String policyReferences = resource.getProperty(PDPConstants.POLICY_REFERENCE);
-            if (policyReferences != null && policyReferences.trim().length() > 0) {
-                dto.setPolicyIdReferences(policyReferences.split(PDPConstants.ATTRIBUTE_SEPARATOR));
-            }
-
-            String policySetReferences = resource.getProperty(PDPConstants.POLICY_SET_REFERENCE);
-            if (policySetReferences != null && policySetReferences.trim().length() > 0) {
-                dto.setPolicySetIdReferences(policySetReferences.split(PDPConstants.ATTRIBUTE_SEPARATOR));
-            }
-            //read policy meta data that is used for basic policy editor
-            dto.setPolicyEditor(resource.getProperty(PDPConstants.POLICY_EDITOR_TYPE));
-            String basicPolicyEditorMetaDataAmount = resource.getProperty(PDPConstants.
-                    BASIC_POLICY_EDITOR_META_DATA_AMOUNT);
-            if (basicPolicyEditorMetaDataAmount != null) {
-                int amount = Integer.parseInt(basicPolicyEditorMetaDataAmount);
-                String[] basicPolicyEditorMetaData = new String[amount];
-                for (int i = 0; i < amount; i++) {
-                    basicPolicyEditorMetaData[i] = resource.
-                            getProperty(PDPConstants.BASIC_POLICY_EDITOR_META_DATA + i);
-                }
-                dto.setPolicyEditorData(basicPolicyEditorMetaData);
-            }
-            PolicyAttributeBuilder policyAttributeBuilder = new PolicyAttributeBuilder();
-            dto.setAttributeDTOs(policyAttributeBuilder.
-                    getPolicyMetaDataFromRegistryProperties(resource.getProperties()));
-            return dto;
-        } catch (RegistryException e) {
-            log.error("Error while loading entitlement policy " + policyId + " from PAP policy store", e);
-            throw new EntitlementException("Error while loading entitlement policy " + policyId +
-                    " from PAP policy store");
+        PolicyDTO dto = store.getPAPPolicy(policyId);
+        if (dto == null) {
+            log.error("Policy does not exist in the system with id " + policyId);
+            throw new EntitlementException("Policy does not exist in the system with id " + policyId);
         }
+        return dto;
     }
 
     /**
-     * Checks whether policy is exist for given policy id
+     * Checks whether policy exists for the given policy id
      *
      * @param policyId policy id
      * @return true of false
      */
     public boolean isExistPolicy(String policyId) {
-        Resource resource = null;
+
+        PolicyDTO dto;
         try {
-            resource = store.getPolicy(policyId, PDPConstants.ENTITLEMENT_POLICY_PAP);
-            if (resource != null) {
+            dto = store.getPAPPolicy(policyId);
+            if (dto != null) {
                 return true;
             }
         } catch (EntitlementException e) {
@@ -188,125 +126,42 @@ public class PAPPolicyStoreReader {
 
 
     /**
-     * Reads Light Weight PolicyDTO for given policy id
+     * Reads light weight PolicyDTO for given policy id
      *
      * @param policyId policy id
-     * @return PolicyDTO but don not contains XACML policy and attribute meta data
+     * @return PolicyDTO but does not contain XACML policy and attribute meta data
      * @throws EntitlementException throws, if fails
      */
     public PolicyDTO readLightPolicyDTO(String policyId) throws EntitlementException {
 
-        Resource resource = null;
-        PolicyDTO dto = null;
-        resource = store.getPolicy(policyId, PDPConstants.ENTITLEMENT_POLICY_PAP);
-        if (resource == null) {
+        PolicyDTO dto = store.getPAPPolicy(policyId);
+        if (dto == null) {
             return null;
         }
-        dto = new PolicyDTO();
-        dto.setPolicyId(policyId);
-        String version = resource.getProperty(PDPConstants.POLICY_VERSION);
-        if (version != null) {
-            dto.setVersion(version);
-        }
-        String lastModifiedTime = resource.getProperty(PDPConstants.LAST_MODIFIED_TIME);
-        if (lastModifiedTime != null) {
-            dto.setLastModifiedTime(lastModifiedTime);
-        }
-        String lastModifiedUser = resource.getProperty(PDPConstants.LAST_MODIFIED_USER);
-        if (lastModifiedUser != null) {
-            dto.setLastModifiedUser(lastModifiedUser);
-        }
-        dto.setActive(Boolean.parseBoolean(resource.getProperty(PDPConstants.ACTIVE_POLICY)));
-        String policyOrder = resource.getProperty(PDPConstants.POLICY_ORDER);
-        if (policyOrder != null) {
-            dto.setPolicyOrder(Integer.parseInt(policyOrder));
-        } else {
-            dto.setPolicyOrder(0);
-        }
-        dto.setPolicyType(resource.getProperty(PDPConstants.POLICY_TYPE));
-
-        String policyReferences = resource.getProperty(PDPConstants.POLICY_REFERENCE);
-        if (policyReferences != null && policyReferences.trim().length() > 0) {
-            dto.setPolicyIdReferences(policyReferences.split(PDPConstants.ATTRIBUTE_SEPARATOR));
-        }
-
-        String policySetReferences = resource.getProperty(PDPConstants.POLICY_SET_REFERENCE);
-        if (policySetReferences != null && policySetReferences.trim().length() > 0) {
-            dto.setPolicySetIdReferences(policySetReferences.split(PDPConstants.ATTRIBUTE_SEPARATOR));
-        }
-
-        dto.setPolicyEditor(resource.getProperty(PDPConstants.POLICY_EDITOR_TYPE));
-
+        dto.setPolicy(null);
+        AttributeDTO[] arr = new AttributeDTO[0];
+        dto.setAttributeDTOs(arr);
+        String[] arr2 = new String[0];
+        dto.setPolicyEditorData(arr2);
         return dto;
     }
 
 
     /**
-     * Reads Light Weight PolicyDTO with Attribute meta data for given policy id
+     * Reads Light Weight PolicyDTO with Attribute metadata for given policy id
      *
      * @param policyId policy id
      * @return PolicyDTO but don not contains XACML policy
      * @throws EntitlementException throws, if fails
      */
     public PolicyDTO readMetaDataPolicyDTO(String policyId) throws EntitlementException {
-        Resource resource = null;
-        PolicyDTO dto = null;
 
-        resource = store.getPolicy(policyId, PDPConstants.ENTITLEMENT_POLICY_PAP);
-        if (resource == null) {
+        PolicyDTO dto = store.getPAPPolicy(policyId);
+        if (dto == null) {
             return null;
         }
-        dto = new PolicyDTO();
-        dto.setPolicyId(policyId);
-        dto.setActive(Boolean.parseBoolean(resource.getProperty(PDPConstants.ACTIVE_POLICY)));
-        String policyOrder = resource.getProperty(PDPConstants.POLICY_ORDER);
-        if (policyOrder != null) {
-            dto.setPolicyOrder(Integer.parseInt(policyOrder));
-        } else {
-            dto.setPolicyOrder(0);
-        }
-
-        String version = resource.getProperty(PDPConstants.POLICY_VERSION);
-        if (version != null) {
-            dto.setVersion(version);
-        }
-        String lastModifiedTime = resource.getProperty(PDPConstants.LAST_MODIFIED_TIME);
-        if (lastModifiedTime != null) {
-            dto.setLastModifiedTime(lastModifiedTime);
-        }
-        String lastModifiedUser = resource.getProperty(PDPConstants.LAST_MODIFIED_USER);
-        if (lastModifiedUser != null) {
-            dto.setLastModifiedUser(lastModifiedUser);
-        }
-        dto.setPolicyType(resource.getProperty(PDPConstants.POLICY_TYPE));
-
-        String policyReferences = resource.getProperty(PDPConstants.POLICY_REFERENCE);
-        if (policyReferences != null && policyReferences.trim().length() > 0) {
-            dto.setPolicyIdReferences(policyReferences.split(PDPConstants.ATTRIBUTE_SEPARATOR));
-        }
-
-        String policySetReferences = resource.getProperty(PDPConstants.POLICY_SET_REFERENCE);
-        if (policySetReferences != null && policySetReferences.trim().length() > 0) {
-            dto.setPolicySetIdReferences(policySetReferences.split(PDPConstants.ATTRIBUTE_SEPARATOR));
-        }
-
-        dto.setPolicyEditor(resource.getProperty(PDPConstants.POLICY_EDITOR_TYPE));
-        String basicPolicyEditorMetaDataAmount = resource.getProperty(PDPConstants.
-                BASIC_POLICY_EDITOR_META_DATA_AMOUNT);
-        if (basicPolicyEditorMetaDataAmount != null) {
-            int amount = Integer.parseInt(basicPolicyEditorMetaDataAmount);
-            String[] basicPolicyEditorMetaData = new String[amount];
-            for (int i = 0; i < amount; i++) {
-                basicPolicyEditorMetaData[i] = resource.
-                        getProperty(PDPConstants.BASIC_POLICY_EDITOR_META_DATA + i);
-            }
-            dto.setPolicyEditorData(basicPolicyEditorMetaData);
-        }
-        PolicyAttributeBuilder policyAttributeBuilder = new PolicyAttributeBuilder();
-        dto.setAttributeDTOs(policyAttributeBuilder.
-                getPolicyMetaDataFromRegistryProperties(resource.getProperties()));
+        dto.setPolicy(null);
         return dto;
-
     }
 
     /**
