+++ components/entitlement/org.wso2.carbon.identity.entitlement/src/main/java/org/wso2/carbon/identity/entitlement/policy/PolicyAttributeBuilder.java
@@ -126,6 +126,7 @@ public class PolicyAttributeBuilder {
         return attributeDTOs.toArray(new AttributeDTO[attributeDTOs.size()]);
     }
 
+
     /**
      * This creates the OMElement from the policy xml and create the the meta data for hole policy
      *
