+++ components/entitlement/org.wso2.carbon.identity.entitlement/src/main/java/org/wso2/carbon/identity/entitlement/policy/finder/CarbonPolicyFinder.java
@@ -37,14 +37,14 @@ import org.wso2.carbon.identity.entitlement.PDPConstants;
 import org.wso2.carbon.identity.entitlement.PolicyOrderComparator;
 import org.wso2.carbon.identity.entitlement.cache.PolicyStatus;
 import org.wso2.carbon.identity.entitlement.common.EntitlementConstants;
+import org.wso2.carbon.identity.entitlement.dao.ConfigDAO;
 import org.wso2.carbon.identity.entitlement.dto.PolicyDTO;
 import org.wso2.carbon.identity.entitlement.internal.EntitlementServiceComponent;
+import org.wso2.carbon.identity.entitlement.pap.EntitlementAdminEngine;
 import org.wso2.carbon.identity.entitlement.pdp.EntitlementEngine;
 import org.wso2.carbon.identity.entitlement.policy.PolicyReader;
 import org.wso2.carbon.identity.entitlement.policy.collection.PolicyCollection;
 import org.wso2.carbon.identity.entitlement.policy.collection.SimplePolicyCollection;
-import org.wso2.carbon.identity.entitlement.policy.store.DefaultPolicyDataStore;
-import org.wso2.carbon.identity.entitlement.policy.store.PolicyDataStore;
 
 import java.net.URI;
 import java.net.URISyntaxException;
@@ -143,12 +143,8 @@ public class CarbonPolicyFinder extends org.wso2.balana.finder.PolicyFinderModul
         if (this.finderModules != null && this.finderModules.size() > 0) {
             // find policy combining algorithm.
 
-            // here we can get policy data store by using EntitlementAdminEngine. But we are not
-            // use it here.  As we need not to have a dependant on EntitlementAdminEngine
-            PolicyDataStore policyDataStore;
-            policyDataStore = new DefaultPolicyDataStore();
-
-            policyCombiningAlgorithm = policyDataStore.getGlobalPolicyAlgorithm();
+            ConfigDAO configDAO = EntitlementAdminEngine.getInstance().getConfigDAO();
+            policyCombiningAlgorithm = configDAO.getGlobalPolicyAlgorithm();
 
             tempPolicyCollection.setPolicyCombiningAlgorithm(policyCombiningAlgorithm);
