+++ components/entitlement/org.wso2.carbon.identity.entitlement/src/main/java/org/wso2/carbon/identity/entitlement/policy/finder/registry/RegistryPolicyHandler.java
@@ -19,8 +19,8 @@
 package org.wso2.carbon.identity.entitlement.policy.finder.registry;
 
 import org.wso2.carbon.identity.entitlement.PDPConstants;
+import org.wso2.carbon.identity.entitlement.dao.RegistryPolicyDAOImpl;
 import org.wso2.carbon.identity.entitlement.internal.EntitlementServiceComponent;
-import org.wso2.carbon.identity.entitlement.policy.store.RegistryPolicyStoreManageModule;
 import org.wso2.carbon.registry.core.exceptions.RegistryException;
 import org.wso2.carbon.registry.core.jdbc.handlers.Handler;
 import org.wso2.carbon.registry.core.jdbc.handlers.RequestContext;
@@ -41,7 +41,7 @@ public class RegistryPolicyHandler extends Handler {
             enableRegistryCacheClear = Boolean.parseBoolean(properties.getProperty(PDPConstants.PDP_REGISTRY_LEVEL_POLICY_CACHE_CLEAR));
         }
         if(enableRegistryCacheClear) {
-            RegistryPolicyStoreManageModule.invalidateCache();
+            RegistryPolicyDAOImpl.invalidateCache();
         }
 
 
@@ -57,7 +57,7 @@ public class RegistryPolicyHandler extends Handler {
             enableRegistryCacheClear = Boolean.parseBoolean(properties.getProperty(PDPConstants.PDP_REGISTRY_LEVEL_POLICY_CACHE_CLEAR));
         }
         if(enableRegistryCacheClear) {
-            RegistryPolicyStoreManageModule.invalidateCache();
+            RegistryPolicyDAOImpl.invalidateCache();
         }
 
     }
